Fix a bug where you weren't able to send transactions from a new account.
This was caused due to a new account not having a public key when no transactions have been performed yet. The problem was fixed by checking if the public key received from the fetch account data response was available and if not, use the local account public key stored in the database, which gets derived from the account address on creation.